requirements.txt
streamlit>=1.31.0
pandas>=2.1.0
plotly>=5.18.0
requests>=2.31.0
msal>=1.26.0
python-dotenv>=1.0.1
openpyxl>=3.1.2
xlrd>=2.0.1
